Add a Visual Studio project for building AtsPluginProxy.dll when needed.
For the time being the original DLL copy in the Dependencies folder is left unchanged and will continue to be used for normal OpenBVE builds so as not to bump the C++ runtime requirement.
