+++ AtsPluginProxy/AtsPluginProxy.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AtsPluginProxy", "AtsPluginProxy\AtsPluginProxy.vcxproj", "{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Debug|x64.ActiveCfg = Debug|x64
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Debug|x64.Build.0 = Debug|x64
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Debug|x86.ActiveCfg = Debug|Win32
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Debug|x86.Build.0 = Debug|Win32
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Release|x64.ActiveCfg = Release|x64
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Release|x64.Build.0 = Release|x64
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Release|x86.ActiveCfg = Release|Win32
+		{CAFBE8AA-37DB-40DD-8418-09703BCCE10D}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
