+++ AtsPluginProxy/AtsPluginProxy/stdafx.h
@@ -1,0 +1,11 @@
+// stdafx.h : include file for standard system include files,
+// or project specific include files that are used frequently,
+// but are changed infrequently
+
+#pragma once
+
+#include "targetver.h"
+
+#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
+// Windows header files:
+#include <windows.h>
