+++ AtsPluginProxy/AtsPluginProxy/targetver.h
@@ -1,0 +1,10 @@
+#pragma once
+
+// Including SDKDDKVer.h defines the highest available Windows platform.
+
+// If you wish to build your application for a previous Windows platform, include WinSDKVer.h and
+// set the _WIN32_WINNT macro to the platform you wish to support before including SDKDDKVer.h.
+
+#include <WinSDKVer.h>
+#define _WIN32_WINNT 0x0501 // _WIN32_WINNT_WINXP
+#include <SDKDDKVer.h>
